User Controller refactored
